move .dart sources into fitting subfolders
following the recommended structure from https://stackoverflow.com/a/64109362/9335596 (similar to other existing recommendations)
run code formatter with "Optimize imports" (Ctrl+Shift+Alt+L in IntelliJ)
