+++ lib/main.dart
@@ -1,17 +1,10 @@
-import 'dart:convert';
-import 'dart:developer';
-import 'dart:io';
-import 'dart:typed_data';
-
-import 'package:canteen_mgmt_frontend/dish_service.dart';
-import 'package:canteen_mgmt_frontend/widget_dish_table.dart';
+import 'package:canteen_mgmt_frontend/widgets/widget_dish_table.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:qr_code_scanner/qr_code_scanner.dart';
-import 'package:http/http.dart' as http;
 
-import 'dish.dart';
-import 'qr_demo.dart';
+import 'models/dish.dart';
+import 'screens/qr_demo.dart';
+import 'services/dish_service.dart';
 
 void main() {
   runApp(const MyApp());
