+++ lib/screens/qr_demo.dart
