+++ lib/services/abstract_service.dart
